change default templating engine for markdown files to nunjucks
By default eleventy uses liquid to pre-process markdown files before running them through
the markdown engine. This is a bit confusing since we're making heavy use of nunjucks
so it makes sense to change the default to nunjucks instead of liquid so we can stick with
a consistent templating syntax.

https://www.11ty.io/docs/config/#default-template-engine-for-markdown-files